refactor: split utilities in separate files (#624)
