+++ src/index.js
@@ -1,6 +1,8 @@
 import fs from 'fs';
 import { resolve, relative, dirname, basename, extname } from 'path';
-import { green, red, yellow, white, blue } from 'kleur';
+import camelCase from 'camelcase';
+import escapeStringRegexp from 'escape-string-regexp';
+import { blue } from 'kleur';
 import { map, series } from 'asyncro';
 import glob from 'tiny-glob/sync';
 import autoprefixer from 'autoprefixer';
@@ -13,93 +15,19 @@ import nodeResolve from '@rollup/plugin-node-resolve';
 import { terser } from 'rollup-plugin-terser';
 import alias from '@rollup/plugin-alias';
 import postcss from 'rollup-plugin-postcss';
-import gzipSize from 'gzip-size';
-import brotliSize from 'brotli-size';
-import prettyBytes from 'pretty-bytes';
 import typescript from 'rollup-plugin-typescript2';
 import json from '@rollup/plugin-json';
 import logError from './log-error';
-import { readFile, isDir, isFile, stdout, stderr, isTruthy } from './utils';
-import camelCase from 'camelcase';
-import escapeStringRegexp from 'escape-string-regexp';
-
-const removeScope = name => name.replace(/^@.*\//, '');
-
-// Convert booleans and int define= values to literals.
-// This is more intuitive than `microbundle --define A=1` producing A="1".
-const toReplacementExpression = (value, name) => {
-	// --define A="1",B='true' produces string:
-	const matches = value.match(/^(['"])(.+)\1$/);
-	if (matches) {
-		return [JSON.stringify(matches[2]), name];
-	}
-
-	// --define @assign=Object.assign replaces expressions with expressions:
-	if (name[0] === '@') {
-		return [value, name.substring(1)];
-	}
-
-	// --define A=1,B=true produces int/boolean literal:
-	if (/^(true|false|\d+)$/i.test(value)) {
-		return [value, name];
-	}
-
-	// default: string literal
-	return [JSON.stringify(value), name];
-};
-
-// Normalize Terser options from microbundle's relaxed JSON format (mutates argument in-place)
-function normalizeMinifyOptions(minifyOptions) {
-	const mangle = minifyOptions.mangle || (minifyOptions.mangle = {});
-	let properties = mangle.properties;
-
-	// allow top-level "properties" key to override mangle.properties (including {properties:false}):
-	if (minifyOptions.properties != null) {
-		properties = mangle.properties =
-			minifyOptions.properties &&
-			Object.assign(properties, minifyOptions.properties);
-	}
-
-	// allow previous format ({ mangle:{regex:'^_',reserved:[]} }):
-	if (minifyOptions.regex || minifyOptions.reserved) {
-		if (!properties) properties = mangle.properties = {};
-		properties.regex = properties.regex || minifyOptions.regex;
-		properties.reserved = properties.reserved || minifyOptions.reserved;
-	}
-
-	if (properties) {
-		if (properties.regex) properties.regex = new RegExp(properties.regex);
-		properties.reserved = [].concat(properties.reserved || []);
-	}
-}
-
-// Parses values of the form "$=jQuery,React=react" into key-value object pairs.
-const parseMappingArgument = (globalStrings, processValue) => {
-	const globals = {};
-	globalStrings.split(',').forEach(globalString => {
-		let [key, value] = globalString.split('=');
-		if (processValue) {
-			const r = processValue(value, key);
-			if (r !== undefined) {
-				if (Array.isArray(r)) {
-					[value, key] = r;
-				} else {
-					value = r;
-				}
-			}
-		}
-		globals[key] = value;
-	});
-	return globals;
-};
-
-// Parses values of the form "$=jQuery,React=react" into key-value object pairs.
-const parseMappingArgumentAlias = aliasStrings => {
-	return aliasStrings.split(',').map(str => {
-		let [key, value] = str.split('=');
-		return { find: key, replacement: value };
-	});
-};
+import { isDir, isFile, stdout, isTruthy, removeScope } from './utils';
+import { getSizeInfo } from './lib/compressed-size';
+import { normalizeMinifyOptions } from './lib/terser';
+import {
+	parseAliasArgument,
+	parseMappingArgument,
+	toReplacementExpression,
+} from './lib/option-normalization';
+import { getConfigFromPkgJson, getName } from './lib/package-info';
+import { shouldCssModules, cssModulesConfig } from './lib/css-modules';
 
 // Extensions to use when resolving modules
 const EXTENSIONS = ['.ts', '.tsx', '.js', '.jsx', '.es6', '.es', '.mjs'];
@@ -107,39 +35,6 @@ const EXTENSIONS = ['.ts', '.tsx', '.js', '.jsx', '.es6', '.es', '.mjs'];
 const WATCH_OPTS = {
 	exclude: 'node_modules/**',
 };
-
-// Hoist function because something (rollup?) incorrectly removes it
-function formatSize(size, filename, type, raw) {
-	const pretty = raw ? `${size} B` : prettyBytes(size);
-	const color = size < 5000 ? green : size > 40000 ? red : yellow;
-	const MAGIC_INDENTATION = type === 'br' ? 13 : 10;
-	return `${' '.repeat(MAGIC_INDENTATION - pretty.length)}${color(
-		pretty,
-	)}: ${white(basename(filename))}.${type}`;
-}
-
-async function getSizeInfo(code, filename, raw) {
-	const gzip = formatSize(
-		await gzipSize(code),
-		filename,
-		'gz',
-		raw || code.length < 5000,
-	);
-	let brotli;
-	//wrap brotliSize in try/catch in case brotli is unavailable due to
-	//lower node version
-	try {
-		brotli = formatSize(
-			await brotliSize(code),
-			filename,
-			'br',
-			raw || code.length < 5000,
-		);
-	} catch (e) {
-		return gzip;
-	}
-	return gzip + '\n' + brotli;
-}
 
 export default async function microbundle(inputOptions) {
 	let options = { ...inputOptions };
@@ -186,12 +81,6 @@ export default async function microbundle(inputOptions) {
 
 	options.multipleEntries = options.entries.length > 1;
 
-	// to disable compress you can put in false or 0 but it's a string so our boolean checks won't work
-	options.compress =
-		typeof options.compress !== 'boolean'
-			? options.compress !== 'false' && options.compress !== '0'
-			: options.compress;
-
 	let formats = (options.format || options.formats).split(',');
 	// always compile cjs first if it's there:
 	formats.sort((a, b) => (a === 'cjs' ? -1 : a > b ? 1 : 0));
@@ -211,53 +100,7 @@ export default async function microbundle(inputOptions) {
 	}
 
 	if (options.watch) {
-		const { onStart, onBuild, onError } = options;
-		return new Promise(resolve => {
-			stdout(
-				blue(
-					`Watching source, compiling to ${relative(
-						cwd,
-						dirname(options.output),
-					)}:`,
-				),
-			);
-
-			const watchers = steps.reduce((acc, options) => {
-				acc[options.inputOptions.input] = watch(
-					Object.assign(
-						{
-							output: options.outputOptions,
-							watch: WATCH_OPTS,
-						},
-						options.inputOptions,
-					),
-				).on('event', e => {
-					if (e.code === 'START') {
-						if (typeof onStart === 'function') {
-							onStart(e);
-						}
-					}
-					if (e.code === 'ERROR') {
-						logError(e.error);
-						if (typeof onError === 'function') {
-							onError(e);
-						}
-					}
-					if (e.code === 'END') {
-						options._sizeInfo.then(text => {
-							stdout(`Wrote ${text.trim()}`);
-						});
-						if (typeof onBuild === 'function') {
-							onBuild(e);
-						}
-					}
-				});
-
-				return acc;
-			}, {});
-
-			resolve({ watchers });
-		});
+		return doWatch(options, cwd, steps);
 	}
 
 	let cache;
@@ -274,67 +117,56 @@ export default async function microbundle(inputOptions) {
 		}),
 	);
 
+	const targetDir = relative(cwd, dirname(options.output)) || '.';
+	const banner = blue(`Build "${options.name}" to ${targetDir}:`);
 	return {
-		output:
-			blue(
-				`Build "${options.name}" to ${relative(cwd, dirname(options.output)) ||
-					'.'}:`,
-			) +
-			'\n   ' +
-			out.join('\n   '),
+		output: `${banner}\n   ${out.join('\n   ')}`,
 	};
 }
 
-async function getConfigFromPkgJson(cwd) {
-	try {
-		const pkgJSON = await readFile(resolve(cwd, 'package.json'), 'utf8');
-		const pkg = JSON.parse(pkgJSON);
+function doWatch(options, cwd, steps) {
+	const { onStart, onBuild, onError } = options;
 
-		return {
-			hasPackageJson: true,
-			pkg,
-		};
-	} catch (err) {
-		const pkgName = basename(cwd);
+	return new Promise((resolve, reject) => {
+		const targetDir = relative(cwd, dirname(options.output));
+		stdout(blue(`Watching source, compiling to ${targetDir}:`));
 
-		stderr(
-			// `Warn ${yellow(`no package.json found. Assuming a pkg.name of "${pkgName}".`)}`
-			yellow(
-				`${yellow().inverse(
-					'WARN',
-				)} no package.json found. Assuming a pkg.name of "${pkgName}".`,
-			),
-		);
-
-		let msg = String(err.message || err);
-		if (!msg.match(/ENOENT/)) stderr(`  ${red().dim(msg)}`);
-
-		return { hasPackageJson: false, pkg: { name: pkgName } };
-	}
-}
-
-const safeVariableName = name =>
-	camelCase(
-		removeScope(name)
-			.toLowerCase()
-			.replace(/((^[^a-zA-Z]+)|[^\w.-])|([^a-zA-Z0-9]+$)/g, ''),
-	);
-
-function getName({ name, pkgName, amdName, cwd, hasPackageJson }) {
-	if (!pkgName) {
-		pkgName = basename(cwd);
-		if (hasPackageJson) {
-			stderr(
-				yellow(
-					`${yellow().inverse(
-						'WARN',
-					)} missing package.json "name" field. Assuming "${pkgName}".`,
+		const watchers = steps.reduce((acc, options) => {
+			acc[options.inputOptions.input] = watch(
+				Object.assign(
+					{
+						output: options.outputOptions,
+						watch: WATCH_OPTS,
+					},
+					options.inputOptions,
 				),
-			);
-		}
-	}
+			).on('event', e => {
+				if (e.code === 'START') {
+					if (typeof onStart === 'function') {
+						onStart(e);
+					}
+				}
+				if (e.code === 'ERROR') {
+					logError(e.error);
+					if (typeof onError === 'function') {
+						onError(e);
+					}
+				}
+				if (e.code === 'END') {
+					options._sizeInfo.then(text => {
+						stdout(`Wrote ${text.trim()}`);
+					});
+					if (typeof onBuild === 'function') {
+						onBuild(e);
+					}
+				}
+			});
 
-	return { finalName: name || amdName || safeVariableName(pkgName), pkgName };
+			return acc;
+		}, {});
+
+		resolve({ watchers });
+	});
 }
 
 async function jsOrTs(cwd, filename) {
@@ -471,9 +303,7 @@ function createConfig(options, entry, format, writeMeta) {
 		outputAliases['.'] = './' + basename(options.output);
 	}
 
-	const moduleAliases = options.alias
-		? parseMappingArgumentAlias(options.alias)
-		: [];
+	const moduleAliases = options.alias ? parseAliasArgument(options.alias) : [];
 	const aliasIds = moduleAliases.map(alias => alias.find);
 
 	const peerDeps = Object.keys(pkg.peerDependencies || {});
@@ -760,54 +590,4 @@ function createConfig(options, entry, format, writeMeta) {
 	};
 
 	return config;
-}
-
-function shouldCssModules(options) {
-	const passedInOption = processCssmodulesArgument(options);
-
-	// We should module when my-file.module.css or my-file.css
-	const moduleAllCss = passedInOption === true;
-
-	// We should module when my-file.module.css
-	const allowOnlySuffixModule = passedInOption === null;
-
-	return moduleAllCss || allowOnlySuffixModule;
-}
-
-function cssModulesConfig(options) {
-	const passedInOption = processCssmodulesArgument(options);
-	const isWatchMode = options.watch;
-	const hasPassedInScopeName = !(
-		typeof passedInOption === 'boolean' || passedInOption === null
-	);
-
-	if (shouldCssModules(options) || hasPassedInScopeName) {
-		let generateScopedName = isWatchMode
-			? '_[name]__[local]__[hash:base64:5]'
-			: '_[hash:base64:5]';
-
-		if (hasPassedInScopeName) {
-			generateScopedName = passedInOption; // would be the string from --css-modules "_[hash]".
-		}
-
-		return { generateScopedName };
-	}
-
-	return false;
-}
-
-/*
-This is done becuase if you use the cli default property, you get a primiatve "null" or "false",
-but when using the cli arguments, you always get back strings. This method aims at correcting those
-for both realms. So that both realms _convert_ into primatives.
-*/
-function processCssmodulesArgument(options) {
-	if (options['css-modules'] === 'true' || options['css-modules'] === true)
-		return true;
-	if (options['css-modules'] === 'false' || options['css-modules'] === false)
-		return false;
-	if (options['css-modules'] === 'null' || options['css-modules'] === null)
-		return null;
-
-	return options['css-modules'];
 }
