+++ src/log-error.js
@@ -1,7 +1,7 @@
 import { red, dim } from 'kleur';
 import { stderr } from './utils';
 
-export default function(err) {
+export default function logError(err) {
 	const error = err.error || err;
 	const description = `${error.name ? error.name + ': ' : ''}${error.message ||
 		error}`;
