+++ src/prog.js
@@ -1,5 +1,5 @@
 import sade from 'sade';
-let { version } = require('../package');
+let { version } = require('../package.json');
 
 const toArray = val => (Array.isArray(val) ? val : val == null ? [] : [val]);
 
@@ -10,9 +10,19 @@ export default handler => {
 
 	const cmd = type => (str, opts) => {
 		opts.watch = opts.watch || type === 'watch';
-		opts.compress =
-			opts.compress != null ? opts.compress : opts.target !== 'node';
+
 		opts.entries = toArray(str || opts.entry).concat(opts._);
+
+		if (opts.compress != null) {
+			// Convert `--compress true/false/1/0` to booleans:
+			if (typeof opts.compress !== 'boolean') {
+				opts.compress = opts.compress !== 'false' && opts.compress !== '0';
+			}
+		} else {
+			// the default compress value is `true` for web, `false` for Node:
+			opts.compress = opts.target !== 'node';
+		}
+
 		handler(opts);
 	};
